create a file for each command
