+++ app/models/airforce/air_force_game.py
@@ -1,9 +1,5 @@
-from hashlib import new
-
 from app.models.airforce.air_force_battlefield import Battlefield
 from app.models.airforce.airforce_filters import get_player_plane
-from app.models.airforce.plane import Projectile
-from app.models.airforce.utils import player_have_plane, position_inside_player_field
 
 
 class AirForceGame:
@@ -97,191 +93,3 @@ class AirForceGame:
         elif self.get_player_plane(self.player_b) == []:
             return self.player_a
         return None
-
-
-class JoinGame:
-    air_force_game = None
-    player = None
-
-    def __init__(self, air_force_game, player):
-        self.air_force_game = air_force_game
-        self.player = player
-
-    def execute(self):
-        if self.air_force_game.player_a == "":
-            self.air_force_game.player_a = self.player
-            self.air_force_game.turn = self.player
-        elif self.air_force_game.player_b == "":
-            self.air_force_game.player_b = self.player
-        else:
-            raise Exception("The game is full!")
-
-
-class GetPlayers:
-    air_force_game = None
-
-    def __init__(self, air_force_game):
-        self.air_force_game = air_force_game
-
-    def execute(self):
-        return {
-            "player_a": self.air_force_game.player_a,
-            "player_b": self.air_force_game.player_b,
-        }
-
-
-class ChoosePlane:
-    course = None
-    x = None
-    y = None
-    player = None
-    battlefield = None
-    plane = None
-    air_force_game = None
-
-    def __init__(self, course, x, y, player, plane, air_force_game):
-        self.course = course
-        self.x = x
-        self.y = y
-        self.player = player
-        self.battlefield = air_force_game.battlefield
-        self.plane = plane
-        self.air_force_game = air_force_game
-
-    def execute(self):
-        if not position_inside_player_field(
-            self.battlefield.max_x,
-            self.x,
-            self.course,
-            self.player,
-            self.air_force_game,
-        ):
-            raise Exception("Plane position cant be inside enemy field")
-        if player_have_plane(self.battlefield, self.player):
-            raise Exception("This player already have a plane")
-
-        return self.battlefield.add_new_flying_object(
-            int(self.player), self.plane, int(self.x), int(self.y), int(self.course)
-        )
-
-
-class MovePlane:
-    course = None
-    player = None
-    battlefield = None
-    air_force_game = None
-
-    def __init__(self, course, player, air_force_game):
-        self.course = course
-        self.player = player
-        self.battlefield = air_force_game.battlefield
-        self.air_force_game = air_force_game
-
-    def execute(self):
-        plane = get_player_plane(self.battlefield, self.player)[0]
-        self.battlefield.move(plane, int(self.course))
-
-
-class LaunchProjectile:
-    player = None
-    air_force_game = None
-    battlefield = None
-
-    def __init__(self, player, air_force_game):
-        self.player = player
-        self.air_force_game = air_force_game
-        self.battlefield = air_force_game.battlefield
-
-    def execute(self):
-        plane = get_player_plane(self.battlefield, int(self.player))[0]
-        course = plane.course
-        x = plane.x
-        y = plane.y
-
-        plane_model = plane.flying_obj
-
-        projectile = Projectile.query.filter_by(plane_id=plane_model.id).first()
-
-        if course == 1:
-            y = y - 1
-        elif course == 2:
-            x = x + 1
-        elif course == 3:
-            y = y + 1
-        elif course == 4:
-            x = x - 1
-
-        return self.battlefield.add_new_flying_object(
-            self.player, projectile, x, y, course
-        )
-
-
-class Shoot:
-    player = None
-    battlefield = None
-
-    def __init__(self, course, player, battlefield):
-        self.course = course
-        self.player = player
-        self.battlefield = battlefield
-
-    def execute(self):
-        raise NotImplementedError()
-
-
-class GetBattlefieldStatus:
-    battlefield = None
-    air_force_game = None
-
-    def __init__(self, battlefield, air_force_game):
-        self.battlefield = battlefield
-        self.air_force_game = air_force_game
-
-    def execute(self):
-        if self.air_force_game.game_ended():
-            return {"status": "end", "Winner": self.air_force_game.winner()}
-        return self.battlefield.get_status()
-
-    # def __str__(self):
-    #     self.battlefield if self.battlefield is not None else "None"
-    #     # self.battlefield.flying_obj if self.battlefield is not None else ""
-
-
-class CheckCourse:
-    course: None
-    game: None
-    player: None
-
-    def __init__(self, course, player, game):
-        self.course = course
-        self.game = game
-        self.player = player
-
-    def execute(self):
-        self.game.battlefield.check_course(self.course, self.player)
-
-
-class GameReady:
-    game: None
-
-    def __init__(self, game):
-        self.game = game
-
-    def execute(self):
-        ready = self.game.player_a != "" and self.game.player_b != ""
-        return {"status": ready}
-
-
-class PlayersHavePlane:
-    game: None
-
-    def __init__(self, game):
-        self.game = game
-
-    def execute(self):
-        ready = (
-            get_player_plane(self.game.battlefield, self.game.player_a) != []
-            and get_player_plane(self.game.battlefield, self.game.player_b) != []
-        )
-
-        return {"status": ready}
