+++ app/models/airforce/commands/check_course.py
@@ -1,0 +1,12 @@
+class CheckCourse:
+    course: None
+    game: None
+    player: None
+
+    def __init__(self, course, player, game):
+        self.course = course
+        self.game = game
+        self.player = player
+
+    def execute(self):
+        self.game.battlefield.check_course(self.course, self.player)
